backend/trash2cash/requirements.txt
celery~=5.5.2
channels
channels-redis
Django~=5.2
django-celery-beat~=2.8.0
django-cors-headers
djangorestframework~=3.16.0
djangorestframework-simplejwt
django-timezone-field~=7.1
google-cloud-storage
google-cloud-vision
googlemaps
Pillow
protobuf~=6.30.2
python-dotenv~=1.1.0
pytz~=2024.1
redis>=4.0.0
